refactor(duration): Public Interface
- Make `duration` mod public
- Remove re-export of `Duration` trait (but
it's still public)

BREAKING CHANGE: The `Duration` trait access changed from
`embedded_time::Duration` to `embedded_time::duration::Duration`
